+++ classes/book_class.rb
