+++ classes/rental_class.rb
