+++ classes/teacher_class.rb
